consistent-sampling: rebase on upstream ComposableSampler API
