+++ consistent-sampling/src/test/java/io/opentelemetry/contrib/sampler/consistent/CoinFlipSampler.java
@@ -3,13 +3,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.sampler.consistent56;
+package io.opentelemetry.contrib.sampler.consistent;
 
 import static java.util.Objects.requireNonNull;
 
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.trace.SpanKind;
 import io.opentelemetry.context.Context;
+import io.opentelemetry.sdk.extension.incubator.trace.samplers.ComposableSampler;
+import io.opentelemetry.sdk.extension.incubator.trace.samplers.SamplingIntent;
 import io.opentelemetry.sdk.trace.data.LinkData;
 import java.util.List;
 import java.util.SplittableRandom;
@@ -20,12 +22,12 @@ import javax.annotation.concurrent.Immutable;
  * of its two delegates. Used by unit tests.
  */
 @Immutable
-final class CoinFlipSampler extends ConsistentSampler {
+final class CoinFlipSampler implements ComposableSampler {
 
   private static final SplittableRandom random = new SplittableRandom(0x160a50a2073e17e6L);
 
-  private final Composable samplerA;
-  private final Composable samplerB;
+  private final ComposableSampler samplerA;
+  private final ComposableSampler samplerB;
   private final double probability;
   private final String description;
 
@@ -36,7 +38,7 @@ final class CoinFlipSampler extends ConsistentSampler {
    * @param samplerA the first delegate sampler
    * @param samplerB the second delegate sampler
    */
-  CoinFlipSampler(Composable samplerA, Composable samplerB) {
+  CoinFlipSampler(ComposableSampler samplerA, ComposableSampler samplerB) {
     this(samplerA, samplerB, 0.5);
   }
 
@@ -48,7 +50,7 @@ final class CoinFlipSampler extends ConsistentSampler {
    * @param samplerA the first delegate sampler
    * @param samplerB the second delegate sampler
    */
-  CoinFlipSampler(Composable samplerA, Composable samplerB, double probability) {
+  CoinFlipSampler(ComposableSampler samplerA, ComposableSampler samplerB, double probability) {
     this.samplerA = requireNonNull(samplerA);
     this.samplerB = requireNonNull(samplerB);
     this.probability = probability;
@@ -66,15 +68,18 @@ final class CoinFlipSampler extends ConsistentSampler {
   @Override
   public SamplingIntent getSamplingIntent(
       Context parentContext,
+      String traceId,
       String name,
       SpanKind spanKind,
       Attributes attributes,
       List<LinkData> parentLinks) {
 
     if (random.nextDouble() < probability) {
-      return samplerA.getSamplingIntent(parentContext, name, spanKind, attributes, parentLinks);
+      return samplerA.getSamplingIntent(
+          parentContext, traceId, name, spanKind, attributes, parentLinks);
     } else {
-      return samplerB.getSamplingIntent(parentContext, name, spanKind, attributes, parentLinks);
+      return samplerB.getSamplingIntent(
+          parentContext, traceId, name, spanKind, attributes, parentLinks);
     }
   }
